Rename models to structured_output
To remove confusion between the llm model and the requested structured output
