+++ config.mdl.ini
@@ -24,7 +24,7 @@ OPENAI_PROJECT_ID =
 # platform section stores selected platform settings
 # select: selected platform between openai and vertexai
 # struct: selected structured output. Choices are directories in the
-#         src/paperext/models directory
+#         src/paperext/structured_output directory
 [platform]
 select = openai
 struct = mdl
