+++ src/paperext/structured_output/ai4hcat/model_v1.py
