+++ src/paperext/structured_output/mdl/model_v1.py
