+++ src/paperext/structured_output/mdl/model_v3.py
